backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-multipart==0.0.12
pdfplumber==0.11.4
python-docx==1.1.2
groq==0.11.0
httpx>=0.27.0,<0.28.0
chromadb==0.5.20
python-dotenv==1.0.1
aiofiles==24.1.0
pydantic==2.9.2
